requirements.txt
flask==1.1.1
flasgger==0.9.1
flask-cors==3.0.8
couchdb==1.2
requests==2.13.0
psycopg2-binary==2.8.3
simplejson==3.16.0
python-dateutil==2.8.0
apscheduler==3.6.1
pytz==2019.2
Pillow==6.1.0
fpdf2==2.0.3
boto3==1.9.253
sh==1.12.14